Add unsaved form handler to next/prev buttons
This disallows moving between forms with the next/prev buttons
while a form is unsaved.
The intended effect is to prevent navigating away from a new form
with unsaved changes.
This is potentially an annoying anti-feature if you wanted to browse
the stack of forms to look for information while filling out another
form.