feat: sync across browsers with "checkForms"
app.pages.form.setInterval (previously nothing) now sends a runner
to "checkForms", which is an alias for "openForms" in gs/Main.
app.refresh, upon receiving "checkForms" will first check that the
current context is still single view mode, then check if the form
has unsaved changes: if not then overwrite with new info, otherwise
check the hash and if the hash matches then do nothing, otherwise
alert the user that their form is out of date.