feat: Implement system of flags
Inital commit to mark items as lost. This commit contains large changes
to the item note taking system, and as such may break item notes written
before this commit. This commit will be followed by another that will
allow items/students to be checked in and forms closed when an item is
lost of a student leaves without checking out.