feat: Implement system of flags - Incomplete Checkouts
allows forms to be closed when items are not returned or students don't
checkout by clicking the item or student and marking the item  as "not found"
or the student as "left without checking out". These put a red warning
in the item's checkIn or student's checkOut column and these cases can
be filtered in the archive.

fixes issue #43