feat: Register new Codabar to a student when scanned
When a codabar is scanned and is not registed with a user, prompt the
user to type a netID to associate the codabar with. If the netID is
already associated with a codabar, ask the user if they want to
overwrite.

Throws errors when the netID cannot be found in the sheet, and when a
codabar is associated with a netID in the sheet but not in app.cache
(in the case that another user registered a codabar with that netID
since the current user's last refresh)

fixes issue #70