fix: lookup students by NetID, not name
After failing to create a booking due to lookup failure on a student
name, it highlights the issue that names are not a good ID system.

This patch updates the Booking model to include a CSV list of ids to
use for identification.

The export from the user database was updated to include this extra
field of data.