fix: form date and time inputs UX improvements
Refactors initializing time inputs to own functions and initialzes both inputs.

Adds a "touched" set for form inputs.

If the end time is not "touched", then the user cannot set an invalid time with the start time.
The app will simply make the end time 1 hour from the start time.