fix: prevents early closed form page state
fixes a bug where the closed forms page could be entered early,
but then the app would unexpectedly switch to the open forms.

The app still allows the closed forms page button to lag behind
the open forms,
but it does not allow *early* unlocking of the closed forms button.